1. 기능 개발 - 유저 대기열 토큰 기능을 개발하고, 테스트를 작성한다.

+++ src/main/kotlin/com/hhplus/concert/presentation/request/BalanceRequest.kt
@@ -2,6 +2,6 @@ package com.hhplus.concert.presentation.request
 
 object BalanceRequest {
     data class Recharge(
-        val amount: Int,
+        val amount: Long,
     )
 }
